feat: add some global state and beginning work on moveset page
Added a wrapper around the app that holds a single pokemon state. This
is for when you're making a moveset for a specific pokemon that you
don't have to make another API call. Not sure if this is the right way
to do it, but it works...for now. This commit also has some beginning
work on the moves page with some basic functionality.