cont: further work on move card
This contains more work on the move card -- almost finalized. There are
2 big structs in pkmnMoves.ts and pokemon.ts for mock values that can be used
for local dev to avoid hitting the pokeAPI endpoints. Of note, in the
move card, the tailwind property bg-[#customHex] doesn't work as
intended out of the box (for now), so use of the styles property on the
JSX elements is used for now.