feat: add postgres and checks for the GET pokemon endpoint
This commit contains a lot. It adds postgres with docker compose through
a separate service. It also begins to add support for goose migrations.
NOTE: Still have to integrate it on the startup of the backend service
in compose.yaml. Because this still has yet to be integrated, have to
exec into the postgres docker container to add the tables. Will continue
to add postgres support for the other endpoints. I used the pgx package
for integration with postgres with the backend. To make the database
available to the handlers, I wrapped the handlers in a handler struct as
methods. Not sure if this is idiomatic, but it works for now.

This is a bit messy, but for now, it works. Will need to refactor later.