feat: getting initial auth together and high level state
This is a step to getting auth together, as well as putting together
high level state for the frontend to maintain some amount of state when
clicking the back button. That's done with React's context api.

Auth is going to be a long haul, but this at least hashes the password,
and stores it in the databse. Still need to add a salt and a bunch of
other functionality, like the login feature, and perhaps adding a
session token or any kind of JWT stuff. Lots of research