feat: further frontend and backend signup work, and some renaming
Renamed some files to be named better.

Added some signup and login endpoints in the backend and have that check
the database. On the frontend, we have a signup page that validates that
fields are present and certain criteria is met, and then gets sent to
the backend. Still need to look at JWTs and sessions.

Beginning on error handling -- still needs a lot of work and more
reserach and understanding.