feat: logout, rough teams page, and topbar work
First, completed the logout functionality. Still a TODO is to have the
frontend delete the cookie on the client.
Secondly, a rough teams page was created and now you can see all the
multiple teams a user has saved, although a very rough styling. It's not
perfect, but at least as of right now, it's functional. Need to make
pretty. Some of the backend code is messy. Should fix
Lastly, added logged in name to topbar, and fixed some button and button
functionality.